110% with edits (commit)

+++ src/main/java/com/demo_banking/models/Transact.java
@@ -10,8 +10,8 @@ import java.time.LocalDateTime;
 public class Transact {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int transaction_id;
-    private int account_id;
+    private Long transaction_id;
+    private Long account_id;
     private String transaction_type;
     private double amount;
     private String source;
@@ -36,19 +36,19 @@ public class Transact {
         this.last_name = last_name;
     }
 
-    public int getTransaction_id() {
+    public Long getTransaction_id() {
         return transaction_id;
     }
 
-    public void setTransaction_id(int transaction_id) {
+    public void setTransaction_id(Long transaction_id) {
         this.transaction_id = transaction_id;
     }
 
-    public int getAccount_id() {
+    public Long getAccount_id() {
         return account_id;
     }
 
-    public void setAccount_id(int account_id) {
+    public void setAccount_id(Long account_id) {
         this.account_id = account_id;
     }
